Extract CartItemRowViewModel to own file
